demos/safety-narrative-generator/requirements.txt
snowflake-connector-python[secure-local-storage]>=3.6.0
snowflake-ml-python>=1.5.0
streamlit>=1.32.0
